Update postinstall.sh
Updated Packages.app so the install path is now
/usr/local/bin 
instead of 
/usr/local/netclient

This should mean that the $PATH linking is not needed, but I've added a command for 'WireGuard-go' to this script, even though it is commented out. Supposedly WireGuard-tools ('wg') are no longer required for netclient to work, but the developers are still including the binary...
So I'm adding everything back in, with hopes of glorious reliability, comrade.